requirements.txt
# Production Dependencies
# Core libraries required for the application to run

# ===== Telegram Client =====
telethon==1.34.0  # Pinned: 1.34.0 (latest stable)
# MTProto API client for Telegram

# ===== Environment & Configuration =====
python-dotenv>=1.0.1  # Pinned: >=1.0.1 (latest stable)
# Environment variable management from .env files

# ===== Progress & User Interface =====
tqdm>=4.66.3  # Pinned: >=4.66.3 (with security fixes)
# Progress bars for command line
rich>=13.7.0  # Pinned: >=13.7.0 (latest stable)
# Rich text and beautiful formatting in terminal

# ===== Async File Operations =====
aiofiles>=23.2.1  # Pinned: >=23.2.1 (latest stable)
# Asynchronous file operations

# ===== System Monitoring =====
psutil>=6.1.0,<7.0.0  # Pinned: ~6.1.0 (compatible with telethon)
# System and process utilities

# ===== Audio Processing =====
mutagen>=1.47.0  # Pinned: >=1.47.0 (latest stable)
# Audio metadata extraction and manipulation
pydub>=0.25.1  # Pinned: >=0.25.1 (latest stable)
# Audio file manipulation and conversion

# ===== Database =====
peewee>=3.17.1  # Pinned: >=3.17.1 (with security fixes)
# Lightweight ORM for SQLite

# ===== Security Updates =====
cryptography>=45.0.6  # Pinned: >=45.0.6 (security updates - fixes multiple CVEs)
pydantic>=2.11.7  # Pinned: >=2.11.7 (security updates)
requests>=2.32.5  # Pinned: >=2.32.5 (security updates)
urllib3>=2.5.0  # Pinned: >=2.5.0 (security updates)

# ===== Command Line Interface =====
click==8.1.7  # Pinned: 8.1.7 (latest stable with security fixes)
# CLI framework for command creation

# ===== HTTP Client =====
aiohttp>=3.12.15  # Pinned: >=3.12.15 (security updates - fixes multiple CVEs)
# Async HTTP client for API calls

# ===== Cryptography =====
cryptography==44.0.1  # Pinned: 44.0.1 (latest stable with security fixes)
# Encryption and security utilities

# ===== Data Validation =====
pydantic>=2.11.7,<3.0.0  # Pinned: >=2.11.7 (latest stable with security fixes)
# Data validation using type hints

# ===== Date and Time =====
python-dateutil>=2.8.2
# Powerful extensions to datetime

# ===== String Similarity =====
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
# Fuzzy string matching

# ===== Configuration Files =====
pyyaml>=6.0
# YAML file parsing for configuration

# ===== Logging =====
structlog>=23.1.0
# Structured logging

# ===== Performance =====
uvloop>=0.17.0; sys_platform != "win32"
# High-performance event loop (Unix only)

# ===== File Type Detection =====
python-magic>=0.4.27; sys_platform != "win32"
python-magic-bin>=0.4.14; sys_platform == "win32"
# File type detection by content

# ===== Retry Logic =====
tenacity>=8.2.0
# Retry library with exponential backoff

# ===== System Notifications =====
plyer>=2.1.0
# Platform-independent API for accessing features of the device

# ===== DEVELOPMENT DEPENDENCIES (Install separately) =====
# For development, use: pip install -e ".[dev]"
# All development dependencies are defined in setup.py

# To install development dependencies:
# pip install pytest>=7.3.1
# pip install pytest-asyncio>=0.21.0  
# pip install pytest-cov>=4.1.0
# pip install pytest-mock>=3.11.1
# pip install pytest-xdist>=3.3.1
# pip install pytest-benchmark>=4.0.0
# pip install black>=23.3.0
# pip install isort>=5.12.0
# pip install flake8>=6.0.0
# pip install mypy>=1.3.0
# pip install bandit>=1.7.5
# pip install safety>=2.3.5
# pip install pre-commit>=3.3.0
# pip install twine>=4.0.0
# pip install build>=0.10.0
# pip install wheel>=0.40.0

# For Docker support:
# pip install docker>=6.1.3
# pip install python-multipart>=0.0.6